feat: Add logic for fetch wallet IFO data
